Migrate `context-menu.test.js`
